changes names and componentes
