+++ code/src/App.js
@@ -1,20 +1,21 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { combineReducers, configureStore } from '@reduxjs/toolkit'
-import quotes from 'reducers/quotes'
+import games from 'reducers/games'
 import StartScreen from 'components/StartScreen'
+import PlayerInput from 'components/PlayerInput'
 // import AuthorInputSearch from 'components/AuthorInputSearch'
 
 export const App = () => {
     const reducer = combineReducers({
-      quotes: quotes.reducer
+      games: games.reducer
     })
 
     const store = configureStore({ reducer })
   return (
     <Provider store={store}>
       <StartScreen />
-      {/* <AuthorInputSearch /> */}
+      <PlayerInput />
     </Provider>
   )
 }
