+++ code/src/components/StartScreen.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import Quotes from './Quotes'
-import AuthorInputSearch from './AuthorInputSearch'
+import GameBoard from './GameBoard'
+import PlayerInput from './PlayerInput'
 
 
 const StartScreen = () => {
-    const quote = useSelector((store) => store.quotes.quote);
+    const game = useSelector((store) => store.games.game);
 
-    return <div>{quote === '' ? <AuthorInputSearch/> : <Quotes/>}</div>
+    return <div>{game === '' ? <PlayerInput/> : <GameBoard/>}</div>
 }
 
 export default StartScreen
